[ContactStructuralMechanicsApplication] Moving several tests to nighty
Basically that
