+++ applications/ContactStructuralMechanicsApplication/tests/ValidationTests.py
@@ -33,6 +33,9 @@ class ALMMultiLayerContactTest(TestFactory):
 class ALMSelfContactContactTest(TestFactory):
     file_name = "ALM_frictionless_contact_test_3D/self_contact_test"
 
+class ExplicitPenaltyThreeDSimplestPatchMatchingTestContact(TestFactory):
+    file_name = "penalty_frictionless_contact_test_3D/explicit_3D_contact_simplest_patch_matching_test"
+
 class ComponentsALMTaylorPatchDynamicTestContact(TestFactory):
     frictionless_by_components = True
     file_name = "ALM_frictionless_contact_test_2D/taylor_patch_dynamic_test"
